+++ todoman/model.py
@@ -543,7 +543,7 @@ class Cache:
 
         return rv
 
-    def todos(self, all=False, lists=[], priority=None, location='',
+    def todos(self, all=False, lists=[], urgent=False, priority=None, location='',
               category='', grep='', sort=[], reverse=True, due=None):
         """
         Returns filtered cached todos, in a specified order.
@@ -557,6 +557,7 @@ class Cache:
 
         :param bool all: If true, also return completed todos.
         :param list lists: Only return todos for these lists.
+        :param bool urgent: Only return urgent todos.
         :param str location: Only return todos with a location containing this
             string.
         :param str category: Only return todos with a category containing this
@@ -586,11 +587,10 @@ class Cache:
             q = ', '.join(['?'] * len(lists))
             extra_where.append('AND files.list_name IN ({})'.format(q))
             params.extend(lists)
-        if priority:
-            if priority == 'medium' or priority == '!low':
-                extra_where.append('AND priority = 5')
-            elif priority == 'high':
-                extra_where.append('AND priority = 1')
+        if urgent:
+            extra_where.append('AND priority = 9')
+        elif priority:
+            extra_where.append('AND priority = ' + priority)
         if location:
             extra_where.append('AND location LIKE ?')
             params.append('%{}%'.format(location))